Many things
Off the top of my head:
 - Refactoring
 - Made things work better internally
 - BootstrapHeapBox doesn't leak anymore
 - Round robin concurrent physmem allocator (woo buzzwords)
 - Removed dead code
